refactor(ui): module layout, split types, colocate Sidebar
- Organize src: api/, store/, types barrel, inbox/, lib (utils+queryClient)
- Feature components under chat/, channels/, agents/ (profile+activity), tasks/
- Pages: MainPanel, TabBar; Sidebar/ with sidebarChannels filter
- Domain types in colocated types.ts; types/index re-exports for api/store
- transport/types for RealtimeMessage; lowercase channels/ in git
- Remove prototype/demo components; update AGENTS.md

Made-with: Cursor
