+++ ui/src/components/agents/activity/ActivityPanel.tsx
@@ -23,7 +23,7 @@ import {
   ChevronUp,
 } from 'lucide-react'
 import { getAgentActivityLog } from '../../../api'
-import type { ActivityLogEntry } from '../../../types'
+import type { ActivityLogEntry } from '../types'
 import './ActivityPanel.css'
 
 interface Props {
