+++ ui/src/components/channels/ChannelMembersPanel.tsx
@@ -1,7 +1,8 @@
 import { useMemo, useState } from 'react'
 import { UserPlus, X } from 'lucide-react'
 import { inviteChannelMember } from '../../api'
-import type { AgentInfo, ChannelMemberInfo, HumanInfo } from '../../types'
+import type { AgentInfo } from '../agents/types'
+import type { ChannelMemberInfo, HumanInfo } from './types'
 import { Dialog, DialogContent, DialogHeader, DialogFooter, DialogTitle, DialogDescription, DialogClose } from '@/components/ui/dialog'
 import { Button } from '@/components/ui/button'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
