+++ ui/src/components/tasks/TasksPanel.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useApp } from '../../store'
 import { useTasks } from '../../hooks/useTasks'
 import { createTasks, updateTaskStatus, claimTasks } from '../../api'
-import type { TaskInfo, TaskStatus } from '../../types'
+import type { TaskInfo, TaskStatus } from './types'
 import { FormError } from '@/components/ui/form'
 import './TasksPanel.css'
