+++ ui/src/components/tasks/types.ts
@@ -1,0 +1,18 @@
+// ── Task board ──
+
+export type TaskStatus = 'todo' | 'in_progress' | 'in_review' | 'done'
+
+export interface TaskInfo {
+  id?: string
+  taskNumber: number
+  title: string
+  status: TaskStatus
+  channelId?: string
+  claimedByName?: string
+  createdByName?: string
+  createdAt?: string
+}
+
+export interface TasksResponse {
+  tasks: TaskInfo[]
+}
