+++ ui/src/pages/Sidebar/Sidebar.tsx
@@ -1,11 +1,12 @@
 import { useEffect, useRef, useState } from 'react'
 import { ChevronDown, Ellipsis, Pencil, Plus, Settings2, Sparkles, Trash2, Users } from 'lucide-react'
-import { useApp } from '../store'
-import type { AgentInfo, ChannelInfo } from '../types'
-import { isVisibleSidebarChannel } from '../sidebarChannels'
-import { CreateAgentModal } from '../components/agents/CreateAgentModal'
-import { CreateChannelModal } from '../components/channels/CreateChannelModal'
-import { DeleteChannelModal, EditChannelModal } from '../components/channels/EditChannelModal'
+import { useApp } from '../../store'
+import type { AgentInfo } from '../../components/agents/types'
+import type { ChannelInfo } from '../../components/channels/types'
+import { isVisibleSidebarChannel } from './sidebarChannels'
+import { CreateAgentModal } from '../../components/agents/CreateAgentModal'
+import { CreateChannelModal } from '../../components/channels/CreateChannelModal'
+import { DeleteChannelModal, EditChannelModal } from '../../components/channels/EditChannelModal'
 import './Sidebar.css'
 
 function agentColor(name: string): string {
