+++ ui/src/pages/Sidebar/index.tsx
@@ -1,0 +1,1 @@
+export { Sidebar } from './Sidebar'
