+++ ui/src/pages/Sidebar/sidebarChannels.ts
@@ -1,4 +1,4 @@
-import type { ChannelInfo } from './types'
+import type { ChannelInfo } from '../../components/channels/types'
 
 export function isVisibleSidebarChannel(channel: ChannelInfo): boolean {
   return channel.joined || channel.channel_type === 'team'
