+++ ui/src/transport/types.ts
@@ -1,0 +1,14 @@
+// ── Realtime WebSocket message envelope ──
+
+import type { StreamEvent } from '../components/chat/types'
+
+export type RealtimeMessage =
+  | {
+      type: 'event'
+      event: StreamEvent
+    }
+  | {
+      type: 'error'
+      code: string
+      message: string
+    }
